将复读，排行榜，历史排行榜，状态整合在一起
