+++ src/plugins/repeat/repeat.py
@@ -1,24 +1,22 @@
-""" 复读插件
+""" 复读
 """
 import re
 import secrets
 from datetime import datetime, timedelta
 
-from nonebot import (CommandSession, IntentCommand, NLPSession, logger,
-                     on_command, on_natural_language, on_notice, permission)
+from nonebot import CommandSession, logger
 
-from coolqbot import PluginData
-
+from . import DATA
 from .recorder import recorder
 
-DATA = PluginData('repeat', config=True)
 # 复读概率
-REPEAT_RATE = int(DATA.get_config('bot', 'repeat_rate', fallback='10'))
+REPEAT_RATE = int(DATA.get_config('repeat', 'repeat_rate', fallback='10'))
 # 复读间隔
-REPEAT_INTERVAL = int(DATA.get_config('bot', 'repeat_interval', fallback='1'))
+REPEAT_INTERVAL = int(DATA.get_config('repeat', 'repeat_interval', fallback='1'))
 
 
-def is_repeat(session: CommandSession, message):
+def is_repeat(session: CommandSession, message: str):
+    """ 是否复读这个消息 """
     group_id = session.ctx['group_id']
     user_id = session.ctx['sender']['user_id']
     # 只复读指定群内消息
@@ -77,40 +75,3 @@ def is_repeat(session: CommandSession, message):
     recorder.add_repeat_list(user_id, group_id)
 
     return True
-
-
-# privileged 设置为 true，防止之前复读没有结束而导致下一个复读判定失败。
-@on_command('repeat', privileged=True)
-async def repeat(session: CommandSession):
-    """ 人类本质
-    """
-    message = session.state.get('message')
-    if is_repeat(session, message):
-        await session.send(message)
-
-
-@on_command('repeat_sign')
-async def repeat_sign(session: CommandSession):
-    """ 复读签到（电脑上没法看手机签到内容）
-    """
-    if session.ctx['group_id'] in session.bot.config.GROUP_ID:
-        title = re.findall(r'title=(\w+\s?\w+)', session.state.get('message'))
-        await session.send(f'今天的运势是{title[0]}', at_sender=True)
-
-
-@on_natural_language(only_to_me=False, permission=permission.GROUP)
-async def _(session: NLPSession):
-    # 只复读群消息，与没有对机器人说的话
-    if not session.ctx['to_me']:
-        # 以置信度 60.0 返回 repeat 命令
-        # 确保任何消息都在且仅在其它自然语言处理器无法理解的时候使用 repeat 命令
-        return IntentCommand(60.0, 'repeat', args={'message': session.msg})
-
-
-@on_natural_language(only_to_me=False)
-async def _(session: NLPSession):
-    match = re.match(r'^\[CQ:sign(.+)\]$', session.msg)
-    if match:
-        return IntentCommand(
-            90.0, 'repeat_sign', args={'message': session.msg}
-        )
