+++ src/plugins/repeat/status.py
@@ -1,19 +1,17 @@
-""" 运行状态插件
+""" 运行状态
 """
 import re
 from datetime import datetime
 
+import aiocqhttp
 from dateutil.relativedelta import relativedelta
-from nonebot import (CommandSession, get_bot, logger, on_command, permission,
-                     scheduler)
+from nonebot import (CommandSession, get_bot, logger, on_websocket_connect)
 
 from .recorder import recorder
 
 
-@on_command(
-    'status', aliases={'状态'}, only_to_me=False, permission=permission.GROUP
-)
-async def status(session: CommandSession):
+async def get_status(session: CommandSession):
+    """ 获取状态 """
     group_id = session.ctx['group_id']
     str_data = f'近十分钟群内聊天数量是 {recorder.message_number(10, group_id)} 条'
 
@@ -59,36 +57,19 @@ def get_total_number(record_list):
     return num
 
 
-@scheduler.scheduled_job('interval', seconds=5, id='coolq_status')
-async def coolq_status():
-    """ 检查酷Q状态
-
-    每5秒检查一次状态，并记录
-    """
-    try:
-        msg = await get_bot().get_status()
-        recorder.coolq_status = msg['good']
-    except:
-        logger.debug('当前无法获取酷Q状态')
-
-
-@scheduler.scheduled_job('interval', seconds=5, id='start_message')
-async def check_status():
-    """ 检测是否需要发送问好信息
-    """
-    if recorder.coolq_status and not recorder.send_hello:
-        hello_str = get_message()
-        for group_id in get_bot().config.GROUP_ID:
-            await get_bot().send_msg(
-                message_type='group', group_id=group_id, message=hello_str
-            )
-        recorder.send_hello = True
-        logger.info('发送首次启动的问好信息')
+@on_websocket_connect
+async def check_status(event: aiocqhttp.Event):
+    """ 启动时发送问好信息 """
+    hello_str = get_message()
+    for group_id in get_bot().config.GROUP_ID:
+        await get_bot().send_msg(
+            message_type='group', group_id=group_id, message=hello_str
+        )
+    logger.info('发送首次启动的问好信息')
 
 
 def get_message():
-    """ 根据当前时间返回对应消息
-    """
+    """ 根据当前时间返回对应消息 """
     hour = datetime.now().hour
 
     if hour > 18 or hour < 6:
